\name{cvplotbestk}
\alias{cvplotbestk}
\title{Plot cv performance on training and test set}
\usage{
cvplotbestk(pdat, title, file)
}
\arguments{
  \item{pdat}{output from \code{\link{cvcreatepdatbestk}}}

  \item{title}{plot title}

  \item{file}{output file. If NULL no save}
}
\description{
Plot cv performance on training and test set
}

