\docType{data}
\name{ktsp}
\alias{ktsp}
\title{Dataset for KTSP package}
\format{A matrix with 1000 genes and 50 samples, binary labels}
\description{
A test dataset for KTSP package taken from the tspair
package by Jeff Leek. The dataset consists of the
expression matrix dat with rows as genes and samples as
columns. The labels are a binary variable named grp.
}
\details{
The data is taken from the tspair package by Jeff Leek.
}
\examples{
data(ktspdata)
}
\references{
Leek, Jeffrey T. "The tspair package for finding top
scoring pair classifiers in R." Bioinformatics 25.9 (2009):
1203-1204.
}
\keyword{datasets}

