\name{tiedRank}
\alias{tiedRank}
\title{tiedRank(x) applies rank to the columns of the matrix x.}
\usage{
tiedRank(x)
}
\arguments{
  \item{x}{a data matrix}
}
\value{
x a rank matrix
}
\description{
tiedRank(x) applies rank to the columns of the matrix x. It
is used to calculate the secondary score.
}
\keyword{rank}

